update param for API filter place

+++ entities/place.go
@@ -11,8 +11,8 @@ type Place struct {
 	Address       string  `json:"address" gorm:"column:address"`
 	Capacity      int     `json:"capacity" gorm:"column:capacity"`
 	Cover         string  `json:"cover" gorm:"column:cover"`
-	Lat           float32 `json:"lat" gorm:"column:lat"`
-	Lng           float32 `json:"lng" gorm:"column:lng"`
+	Lat           float64 `json:"lat" gorm:"column:lat"`
+	Lng           float64 `json:"lng" gorm:"column:lng"`
 	Country       string  `json:"country" gorm:"column:country"`
 	State         string  `json:"state" gorm:"column:state"`
 	District      string  `json:"district" gorm:"column:district"`
