+++ provider/googlemap/model.go
@@ -30,8 +30,8 @@ type AddressComponent struct {
 }
 
 type Location struct {
-	Lat float32 `json:"lat"`
-	Lng float32 `json:"lng"`
+	Lat float64 `json:"lat"`
+	Lng float64 `json:"lng"`
 }
 
 type Geometry struct {
